Refactor baseline profile generating scenarios
