+++ benchmark/src/main/kotlin/io/getstream/video/android/benchmark/BaselineProfileGenerator.kt
@@ -19,11 +19,6 @@ package io.getstream.video.android.benchmark
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.benchmark.macro.junit4.BaselineProfileRule
-import androidx.test.uiautomator.By
-import androidx.test.uiautomator.BySelector
-import androidx.test.uiautomator.UiDevice
-import androidx.test.uiautomator.UiObject2
-import androidx.test.uiautomator.Until
 import org.junit.Rule
 import org.junit.Test
 
@@ -43,76 +38,6 @@ internal class BaselineProfileGenerator {
             startActivityAndWait()
             device.waitForIdle()
 
-            // -------------
-            // Authenticate
-            // -------------
-            device.authenticateAndNavigateToHome()
-
-            // -------------
-            // JoinCall
-            // -------------
-            device.navigateFromJoinCallToLobby()
-
-            // -------------
-            // Lobby
-            // -------------
-            device.navigateFromLobbyToCall()
-
-            // -------------
-            // Call
-            // -------------
-            device.testCall()
+            dogfoodingScenarios()
         }
-}
-
-internal fun UiDevice.authenticateAndNavigateToHome() {
-    wait(Until.hasObject(By.res("authenticate")), 5_000)
-
-    // Click the Authenticate button and login.
-    waitForObject(By.res("authenticate"))?.click()
-
-    waitForIdle()
-}
-
-internal fun UiDevice.navigateFromJoinCallToLobby() {
-    wait(Until.hasObject(By.res("start_new_call")), 5_000)
-
-    // wait for the Join Call button and navigate to the lobby screen by clicking.
-    waitForObject(By.res("start_new_call"))?.click()
-
-    waitForIdle()
-}
-
-internal fun UiDevice.navigateFromLobbyToCall() {
-    waitForObject(By.text(getPermissionText()), 5_000)?.click()
-    waitForObject(By.text(getPermissionText()), 5_000)?.click()
-
-    wait(Until.hasObject(By.res("call_lobby")), 15_000)
-    wait(Until.hasObject(By.res("participant_video_renderer")), 15_000)
-
-    // wait for the Start Call button and navigate to the call screen by clicking.
-    waitForObject(By.res("start_call"), 15_000)?.click()
-
-    waitForIdle()
-}
-
-internal fun UiDevice.testCall() {
-    wait(Until.hasObject(By.res("call_content")), 5_000)
-    wait(Until.hasObject(By.res("video_renderer")), 5_000)
-    waitForIdle()
-}
-
-internal fun UiDevice.waitForObject(selector: BySelector, timeout: Long = 5_000): UiObject2? {
-    if (wait(Until.hasObject(selector), timeout)) {
-        return findObject(selector)
-    }
-    return null
-}
-
-internal fun getPermissionText(): String {
-    return when {
-        Build.VERSION.SDK_INT <= 28 -> "ALLOW"
-        Build.VERSION.SDK_INT == 29 -> "Allow only while using the app"
-        else -> "While using the app"
-    }
 }
